+++ src/v1/HttpStatusCode/mustBeHttpStatusCode.spec.ts
@@ -31,20 +31,20 @@
 // ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
 // POSSIBILITY OF SUCH DAMAGE.
 //
-import { OnError } from "@ganbarodigital/ts-on-error/lib/V1";
+import { AnyAppError, OnError } from "@ganbarodigital/ts-lib-error-reporting/lib/v1";
 import { expect } from "chai";
 import { describe } from "mocha";
 
 import { mustBeHttpStatusCode } from "./mustBeHttpStatusCode";
 
 describe("mustBeHttpStatusCode()", () => {
-    const onError: OnError = (reason: symbol, desc: string, extra: object): never => {
-        throw new Error(JSON.stringify(extra));
+    const onError: OnError = (e: AnyAppError): never => {
+        throw new Error(JSON.stringify(e.details.extra));
     };
 
     it("accepts integers in the range 100-599 inclusive", () => {
         for (let inputValue = 100; inputValue < 600; inputValue++) {
-            mustBeHttpStatusCode(inputValue, onError);
+            mustBeHttpStatusCode(inputValue);
         }
     });
 
@@ -67,5 +67,10 @@ describe("mustBeHttpStatusCode()", () => {
             const expectedMessage = "{\"input\":" + inputValue + "}";
             expect(() => mustBeHttpStatusCode(inputValue, onError)).to.throw(expectedMessage);
         }
+    });
+
+    it("has a default error handler", () => {
+        const expectedMessage = "input falls outside the range of a valid HTTP status code";
+        expect(() => mustBeHttpStatusCode(700)).to.throw(expectedMessage);
     });
 });