+++ CHANGELOG.md
@@ -21,6 +21,10 @@ For each release, changes are grouped under these headings:
 
 The following changes have been completed, and will be included in the next tagged release.
 
+### Backwards-Compatibility Breaks
+
+* Updated error callbacks to use `@ganbarodigital/ts-lib-error-reporting`
+
 ## v0.0.2
 
 Released Wednesday, 19th February 2020.